fix : separation of osm_interface from osm plugin * add missing attribute to buildWidget in osm_flutter * add body of dispose in base_map_controller * hide init of base_map_controller
