+++ flutter_osm_interface/lib/src/map_controller/i_base_map_controller.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_osm_interface/flutter_osm_interface.dart';
 import '../types/types.dart';
@@ -16,6 +17,17 @@ abstract class IBaseMapController {
 
   // late ValueNotifier<GeoPoint?> listenerMapLongTapping = ValueNotifier(null);
   // late ValueNotifier<GeoPoint?> listenerMapSingleTapping = ValueNotifier(null);
+  late ValueNotifier<GeoPoint?> _listenerMapLongTapping = ValueNotifier(null);
+  late ValueNotifier<GeoPoint?> _listenerMapSingleTapping = ValueNotifier(null);
+  late ValueNotifier<bool> _listenerMapIsReady = ValueNotifier(false);
+
+  ValueListenable<GeoPoint?> get listenerMapLongTapping =>
+      _listenerMapLongTapping;
+
+  ValueListenable<GeoPoint?> get listenerMapSingleTapping =>
+      _listenerMapSingleTapping;
+
+  ValueListenable<bool> get listenerMapIsReady => _listenerMapIsReady;
 
   IBaseMapController({
     this.initMapWithUserPosition = true,
@@ -24,7 +36,25 @@ abstract class IBaseMapController {
 
   }) : assert(initMapWithUserPosition || initPosition != null);
 
-  void init();
+  //void init();
 
-  void dispose();
+  void dispose(){
+    _listenerMapLongTapping.dispose();
+    _listenerMapSingleTapping.dispose();
+    _listenerMapIsReady.dispose();
+  }
+}
+
+extension setLiteners on BaseMapController {
+  void setValueListenerMapLongTapping(GeoPoint p) {
+    _listenerMapLongTapping.value = p;
+  }
+
+  void setValueListenerMapSingleTapping(GeoPoint p) {
+    _listenerMapSingleTapping.value = p;
+  }
+
+  void setValueListenerMapIsReady(bool isReady) {
+    _listenerMapIsReady.value = isReady;
+  }
 }
