+++ flutter_osm_interface/lib/src/osm_interface.dart
@@ -37,211 +37,211 @@ abstract class OSMPlatform extends PlatformInterface {
 
   void close();
 
-  // Future<void> configureZoomMap(
-  //   int idOSM,
-  //   double initZoom,
-  //   int minZoomLevel,
-  //   int maxZoomLevel,
-  //   double stepZoom,
-  // );
+// Future<void> configureZoomMap(
+//   int idOSM,
+//   double initZoom,
+//   int minZoomLevel,
+//   int maxZoomLevel,
+//   double stepZoom,
+// );
 }
 
 abstract class MobileOSMPlatform extends OSMPlatform {
   Future<void> initMap(
-      int idOSM,
-      GeoPoint point,
-      );
+    int idOSM,
+    GeoPoint point,
+  );
 
   Future<void> currentLocation(
-      int idOSM,
-      );
+    int idOSM,
+  );
 
   Future<GeoPoint> myLocation(
-      int idMap,
-      );
+    int idMap,
+  );
 
   Future<GeoPoint> pickLocation(
-      int idOSM, {
-        GlobalKey? key,
-        String imageURL = "",
-      });
+    int idOSM, {
+    GlobalKey? key,
+    String imageURL = "",
+  });
 
   Future<void> customMarker(
-      int idOSM,
-      GlobalKey? globalKey,
-      );
+    int idOSM,
+    GlobalKey? globalKey,
+  );
 
   Future<void> customUserLocationMarker(
-      int idOSM,
-      GlobalKey personGlobalKey,
-      GlobalKey directionArrowGlobalKey,
-      );
+    int idOSM,
+    GlobalKey personGlobalKey,
+    GlobalKey directionArrowGlobalKey,
+  );
 
   Future<void> setColorRoad(
-      int idOSM,
-      Color color,
-      );
+    int idOSM,
+    Color color,
+  );
 
   Future<void> setMarkersRoad(
-      int idOSM,
-      List<GlobalKey?> keys,
-      );
+    int idOSM,
+    List<GlobalKey?> keys,
+  );
 
   Future<void> enableTracking(
-      int idOSM,
-      );
+    int idOSM,
+  );
 
   Future<void> addPosition(
-      int idOSM,
-      GeoPoint p,
-      );
+    int idOSM,
+    GeoPoint p,
+  );
 
   Future<void> goToPosition(
-      int idOSM,
-      GeoPoint p,
-      );
+    int idOSM,
+    GeoPoint p,
+  );
 
   Future<void> addMarker(
-      int idOSM,
-      GeoPoint p, {
-        GlobalKey? globalKeyIcon,
-      });
+    int idOSM,
+    GeoPoint p, {
+    GlobalKey? globalKeyIcon,
+  });
 
   Future<void> removePosition(
-      int idOSM,
-      GeoPoint p,
-      );
+    int idOSM,
+    GeoPoint p,
+  );
 
   Future<void> removeLastRoad(
-      int idOSM,
-      );
+    int idOSM,
+  );
 
   Future<RoadInfo> drawRoad(
-      int idOSM,
-      GeoPoint start,
-      GeoPoint end, {
-        RoadType roadType = RoadType.car,
-        List<GeoPoint>? interestPoints,
-        RoadOption roadOption,
-      });
+    int idOSM,
+    GeoPoint start,
+    GeoPoint end, {
+    RoadType roadType = RoadType.car,
+    List<GeoPoint>? interestPoints,
+    RoadOption roadOption,
+  });
 
   Future<void> drawCircle(
-      int idOSM,
-      CircleOSM circleOSM,
-      );
+    int idOSM,
+    CircleOSM circleOSM,
+  );
 
   Future<void> removeCircle(
-      int idOSM,
-      String key,
-      );
+    int idOSM,
+    String key,
+  );
 
   Future<void> drawRect(
-      int idOSM,
-      RectOSM rectOSM,
-      );
+    int idOSM,
+    RectOSM rectOSM,
+  );
 
   Future<void> removeRect(
-      int idOSM,
-      String key,
-      );
+    int idOSM,
+    String key,
+  );
 
   Future<void> removeAllRect(
-      int idOSM,
-      );
+    int idOSM,
+  );
 
   Future<void> removeAllCircle(
-      int idOSM,
-      );
+    int idOSM,
+  );
 
   Future<void> removeAllShapes(
-      int idOSM,
-      );
+    int idOSM,
+  );
 
   Future<void> customMarkerStaticPosition(
-      int idOSM,
-      GlobalKey? globalKey,
-      String id,
-      );
+    int idOSM,
+    GlobalKey? globalKey,
+    String id,
+  );
 
   Future<void> staticPosition(
-      int idOSM,
-      List<GeoPoint> pList,
-      String id,
-      );
+    int idOSM,
+    List<GeoPoint> pList,
+    String id,
+  );
 
   Future<double> getZoom(
-      int idOSM,
-      );
+    int idOSM,
+  );
 
   Future<void> setZoom(
-      int idOSM, {
-        double? zoomLevel,
-        double? stepZoom,
-      });
+    int idOSM, {
+    double? zoomLevel,
+    double? stepZoom,
+  });
 
   Future<void> setStepZoom(
-      int idOSM,
-      int stepZoom,
-      );
+    int idOSM,
+    int stepZoom,
+  );
 
   Future<void> setMinimumZoomLevel(
-      int idOSM,
-      int minZoom,
-      );
+    int idOSM,
+    int minZoom,
+  );
 
   Future<void> setMaximumZoomLevel(
-      int idOSM,
-      int maxZoom,
-      );
+    int idOSM,
+    int maxZoom,
+  );
 
   Future<void> disableTracking(
-      int idOSM,
-      );
+    int idOSM,
+  );
 
   Future<void> visibilityInfoWindow(
-      int idOSM,
-      bool visible,
-      );
+    int idOSM,
+    bool visible,
+  );
 
   Future<void> advancedPositionPicker(
-      int idOSM,
-      );
+    int idOSM,
+  );
 
   Future<GeoPoint> getPositionOnlyAdvancedPositionPicker(
-      int idOSM,
-      );
+    int idOSM,
+  );
 
   Future<GeoPoint> selectAdvancedPositionPicker(
-      int idOSM,
-      );
+    int idOSM,
+  );
 
   Future<void> cancelAdvancedPositionPicker(
-      int idOSM,
-      );
+    int idOSM,
+  );
 
   Future<void> drawRoadManually(
-      int idOSM,
-      List<GeoPoint> road,
-      Color roadColor,
-      double width,
-      );
+    int idOSM,
+    List<GeoPoint> road,
+    Color roadColor,
+    double width,
+  );
 
   Future<void> mapRotation(
-      int idOSM,
-      double degree,
-      );
+    int idOSM,
+    double degree,
+  );
 
   Future<void> customAdvancedPickerMarker(
-      int idMap,
-      GlobalKey key,
-      );
+    int idMap,
+    GlobalKey key,
+  );
 
   Future<void> limitArea(
-      int idOSM,
-      BoundingBox box,
-      );
+    int idOSM,
+    BoundingBox box,
+  );
 
   Future<void> removeLimitArea(
-      int idOSM,
-      );
+    int idOSM,
+  );
 }
